+++ src/layers/cip/EPath/index.js
@@ -87,7 +87,6 @@ class EPath {
           break;
         case 5:
         case 6:
-          // throw new Error(`Use datatypes/decoding instead of EPath`);
           offset = Segments.DataType.Decode(segmentCode, buffer, offset, padded, val => segments.push(val));
           break;
         default: